Add usable library for this crate.
